Fixed pie chart to accomodate multiple instances of same category and added remaining allowance indicator

+++ BudgetTracker/budget/models.py
@@ -15,7 +15,7 @@ class Category(models.Model):
 class Budget(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
     categories = models.ManyToManyField(Category)
-    allowance = models.IntegerField()
+    allowance = models.FloatField()
 
     def __str__(self):
         return f'{self.user}, {self.categories.all()}, {self.allowance }'
@@ -25,7 +25,7 @@ class Transaction(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     category = models.ForeignKey(Category, max_length=50, on_delete=models.CASCADE)
     transaction_date = models.DateField()
-    amount_spent = models.IntegerField()
+    amount_spent = models.FloatField()
 
     def __str__(self):
         return f'{self.user}, {self.category}, {self.transaction_date }, {self.amount_spent}'
